pyproject.toml
[project]
authors = []
dependencies = [
    "strawberry-graphql",
    "fastapi",
    "uvicorn",
    "rich>=13.0.1",
    "httpx>=0.23.3",
    "humanize>=4.4.0",
    "sentry-sdk[fastapi]>=1.14.0",
]
description = ""
name = "api"
requires-python = ">=3.7"
version = "0.1.0"

[tool.black]
line-length = 88

[tool.isort]
combine_as_imports = true
combine_star = true
indent = 4
known_pytest = ["pytest"]
known_strawberry = ["strawberry"]
lines_after_imports = 2
profile = "black"
sections = ["FUTURE", "STDLIB", "PYTEST", "THIRDPARTY", "STRAWBERRY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
plugins = ["strawberry.ext.mypy_plugin"]

[tool.pdm]
[tool.pdm.scripts]
dev.cmd = "uvicorn main:app --reload"
dev.env_file = ".env"

start = "uvicorn main:app --host 0.0.0.0 --port 8080"

test.cmd = "pytest"
test.env = {TFL_API_KEY = "bar"}

export-schema.cmd = "strawberry export-schema api.schema"
export-schema.env = {TFL_API_KEY = "bar"}

[tool.pdm.dev-dependencies]
dev = [
  "black>=22.12.0",
  "pytest>=7.2.0",
  "pdbpp>=0.10.3",
]
[build-system]
build-backend = "pdm.pep517.api"
requires = ["pdm-pep517>=1.0.0"]
